refactor(trip): 旅程画面をタスク画面と同じ構成に分割
